+++ app/lib/search/index_simple.dart
@@ -168,6 +168,22 @@ class SimplePackageIndex implements PackageIndex {
       });
     }
 
+    // filter on owners
+    if (query.uploaderOrPublishers != null) {
+      assert(query.uploaderOrPublishers.isNotEmpty);
+
+      packages.removeWhere((package) {
+        final doc = _packages[package];
+        if (doc.publisherId != null) {
+          return !query.uploaderOrPublishers.contains(doc.publisherId);
+        }
+        if (doc.uploaderEmails == null) {
+          return true; // turn this into an error in the future.
+        }
+        return !query.uploaderOrPublishers.any(doc.uploaderEmails.contains);
+      });
+    }
+
     // filter on publisher
     if (query.publisherId != null || query.parsedQuery.publisher != null) {
       final publisherId = query.publisherId ?? query.parsedQuery.publisher;