+++ app/lib/search/search_service.g.dart
@@ -31,6 +31,8 @@ PackageDocument _$PackageDocumentFromJson(Map<String, dynamic> json) {
     ),
     publisherId: json['publisherId'] as String,
     emails: (json['emails'] as List)?.map((e) => e as String)?.toList(),
+    uploaderEmails:
+        (json['uploaderEmails'] as List)?.map((e) => e as String)?.toList(),
     apiDocPages: (json['apiDocPages'] as List)
         ?.map((e) =>
             e == null ? null : ApiDocPage.fromJson(e as Map<String, dynamic>))
@@ -60,6 +62,7 @@ Map<String, dynamic> _$PackageDocumentToJson(PackageDocument instance) =>
       'dependencies': instance.dependencies,
       'publisherId': instance.publisherId,
       'emails': instance.emails,
+      'uploaderEmails': instance.uploaderEmails,
       'apiDocPages': instance.apiDocPages,
       'timestamp': instance.timestamp?.toIso8601String(),
     };